requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
email-validator==2.1.0